+++ src/Discord.Net.Core/Entities/Messages/ISystemMessage.cs
@@ -3,7 +3,7 @@ namespace Discord
     /// <summary>
     ///     Represents a generic message sent by the system.
     /// </summary>
-    public interface ISystemMessage : IReactionMessage
+    public interface ISystemMessage : IMessage
     {
     }
 }
