+++ src/Discord.Net.Rest/Entities/Messages/RestMessage.cs
@@ -13,6 +13,7 @@ namespace Discord.Rest
     public abstract class RestMessage : RestEntity<ulong>, IMessage, IUpdateable
     {
         private long _timestampTicks;
+        private ImmutableArray<RestReaction> _reactions = ImmutableArray.Create<RestReaction>();
 
         /// <inheritdoc />
         public IMessageChannel Channel { get; }
@@ -106,6 +107,22 @@ namespace Discord.Rest
                     PartyId = model.Activity.Value.PartyId.GetValueOrDefault()
                 };
             }
+
+            if (model.Reactions.IsSpecified)
+            {
+                var value = model.Reactions.Value;
+                if (value.Length > 0)
+                {
+                    var reactions = ImmutableArray.CreateBuilder<RestReaction>(value.Length);
+                    for (int i = 0; i < value.Length; i++)
+                        reactions.Add(RestReaction.Create(value[i]));
+                    _reactions = reactions.ToImmutable();
+                }
+                else
+                    _reactions = ImmutableArray.Create<RestReaction>();
+            }
+            else
+                _reactions = ImmutableArray.Create<RestReaction>();
         }
 
         /// <inheritdoc />
@@ -135,5 +152,24 @@ namespace Discord.Rest
         IReadOnlyCollection<IEmbed> IMessage.Embeds => Embeds;
         /// <inheritdoc />
         IReadOnlyCollection<ulong> IMessage.MentionedUserIds => MentionedUsers.Select(x => x.Id).ToImmutableArray();
+       
+        /// <inheritdoc />
+        public IReadOnlyDictionary<IEmote, ReactionMetadata> Reactions => _reactions.ToDictionary(x => x.Emote, x => new ReactionMetadata { ReactionCount = x.Count, IsMe = x.Me });
+
+        /// <inheritdoc />
+        public Task AddReactionAsync(IEmote emote, RequestOptions options = null)
+            => MessageHelper.AddReactionAsync(this, emote, Discord, options);
+        /// <inheritdoc />
+        public Task RemoveReactionAsync(IEmote emote, IUser user, RequestOptions options = null)
+            => MessageHelper.RemoveReactionAsync(this, user.Id, emote, Discord, options);
+        /// <inheritdoc />
+        public Task RemoveReactionAsync(IEmote emote, ulong userId, RequestOptions options = null)
+            => MessageHelper.RemoveReactionAsync(this, userId, emote, Discord, options);
+        /// <inheritdoc />
+        public Task RemoveAllReactionsAsync(RequestOptions options = null)
+            => MessageHelper.RemoveAllReactionsAsync(this, Discord, options);
+        /// <inheritdoc />
+        public IAsyncEnumerable<IReadOnlyCollection<IUser>> GetReactionUsersAsync(IEmote emote, int limit, RequestOptions options = null)
+            => MessageHelper.GetReactionUsersAsync(this, emote, limit, Discord, options);
     }
 }
