+++ src/Discord.Net.Rest/Entities/Messages/RestUserMessage.cs
@@ -11,7 +11,7 @@ namespace Discord.Rest
     ///     Represents a REST-based message sent by a user.
     /// </summary>
     [DebuggerDisplay(@"{DebuggerDisplay,nq}")]
-    public class RestUserMessage : RestReactionMessage, IUserMessage
+    public class RestUserMessage : RestMessage, IUserMessage
     {
         private bool _isMentioningEveryone, _isTTS, _isPinned, _isSuppressed;
         private long? _editedTimestampTicks;
